Modifiqué la estructura modular de los archivos .js al igual que la vinculación de todas las páginas .html
