+++ Frontend/pages/products.html
@@ -573,7 +573,7 @@
                 </div>
             </div>
         </footer>
-        <script type="module" src="../src/themeModeProducts.js"></script>
+        <script type="module" src="../src/main.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
     </body>
 </html>
